@@ -1,49 +1,33 @@
-[build-system]
-requires = ["flit"]
-build-backend = "flit.buildapi"
+[tool.poetry]
+name = "django-websocket-notification"
+version = "2.1.2"
+description = "A Django package to handle notifications using Django Channels and WebSockets."
+readme = "README.rst"
+authors = ["Marcos Gabarda <hey@marcosgabarda.com>"]
+license = "MIT"
+homepage = "https://github.com/marcosgabarda/django-websocket-notifications"
+packages = [
+    { include = "options" },
+]
 
-[tool.flit.metadata]
-dist-name = "django-simple-options"
-module = "options"
-author = "Marcos Gabarda"
-author-email = "hey@marcosgabarda.com"
-home-page = "https://github.com/marcosgabarda/django-simple-options"
 classifiers = [
-        "Environment :: Web Environment",
-        "Framework :: Django",
-        "Intended Audience :: Developers",
-        "License :: OSI Approved :: MIT License",
-        "Operating System :: OS Independent",
-        "Programming Language :: Python :: 3",
-        "Topic :: Utilities",
+    "Environment :: Web Environment",
+    "Framework :: Django",
+    "License :: OSI Approved :: MIT License",
 ]
-requires = [
-    "django"
- ]
-description-file = "README.rst"
-requires-python=">=3.6"
+[tool.poetry.dependencies]
+python = "^3.7"
+django = "^3.0.7"
+djangorestframework = "^3.11.0"
 
-[tool.flit.metadata.requires-extra]
-test = [
-    "pytest",
-    "pytest-django",
-    "pytest-cov",
-    "factory_boy",
-]
-doc = ["sphinx"]
+[tool.poetry.dev-dependencies]
+pylint = "^2.5.3"
+pytest = "^5.4.3"
+black = "^19.10b0"
+factory_boy = "^2.12.0"
+django-test-plus = "^1.4.0"
+pytest-django = "^3.9.0"
 
-[tool.tox]
-legacy_tox_ini = """
-[tox]
-skipsdist=True
-envlist = py36,py37
-
-[testenv]
-deps =
-    django
-    pytest
-    pytest-django
-    pytest-cov
-    factory_boy
-commands = pytest
-"""
+[build-system]
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
