@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.db import models
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from options.constants import STR, TYPE_CHOICES, CONVERTER, FILE
 from options.helpers import convert_value, UploadToDir
